requirements.txt
absl-py==1.0.0
astunparse==1.6.3
av==8.0.3
Bottleneck==1.3.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
einops==0.3.2
flatbuffers==2.0
fvcore==0.1.5.post20211023
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.5.0
idna==3.3
iopath==0.1.9
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
natsort==8.0.0
numexpr @ file:///tmp/build/80754af9/numexpr_1618856167419/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095647912/work
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.4.58
opt-einsum==3.3.0
pandas==1.3.4
Pillow==8.4.0
portalocker==2.3.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six @ file:///tmp/build/80754af9/six_1623709665295/work
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
yacs==0.1.8
